Additional testing on sqliter driver and resources
